MOTECH-2296 Task retries for event-based actions
This commit enhances the task retries for the event-based
task actions. Default event system retries are no longer
invoked for task actions. The event module allows to register
event callbacks now, that will be invoked when the event
is handled. Moreover, the Task module now keeps the track
of the task executions.