MOTECH-2296 Ability to register callbacks to MotechEvents
This commit adds the ability to register callback services, that
will be invoked when the execution of a MotechEvent is finished.
The service must implement the EventCallbackService and provide
a callback name, that must be also included in the MotechEvent.
The interface provides two methods - one for success callbacks
and another one for failure callbacks.