MOTECH-2296 Task retries for event-based actions (#536)
* MOTECH-2296 Task retries for event-based actions

This commit enhances the task retries for the event-based
task actions. Default event system retries are no longer
invoked for task actions. The event module allows to register
event callbacks now, that will be invoked when the event
is handled. Moreover, the Task module now keeps the track
of the task executions.

* MOTECH-2296 Review comments incorporation

* MOTECH-2296 Small refactor

* MOTECH-2296 Test fixes

* MOTECH-2296 Resolved conflicts and small fixes

* MOTECH-2296: Fixed multiply reply count and bubbles parameters

* MOTECH-2296: Changed condition

* MOTECH-2296: Checkstyle...

* MOTECH-2296: Added trigger params to event - without fixed tests

* MOTECH-2296: Added condition to adding task params only for events